requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mpmath==1.1.0
pymongo==3.10.1
python-dotenv==0.11.0
requests==2.22.0
sympy==1.5.1
urllib3==1.25.8
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
